Adds homeworks

+++ src/gNumericTypes/Overflow.java
@@ -4,7 +4,7 @@ public class Overflow {
 
 	public static void main(String args[])
 	{
-		int n = 1;
+		long n = 1;
 	    for (int i = 0; i < 64; i++)
 	    {
 	    	System.out.print("n: " + n + "\n");
